pyproject.toml
[tool.poetry]
name = "edtech-doubt-resolution"
version = "0.1.0"
description = "YES Doubt-Resolution Platform Backend"
authors = ["Developer"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.135.1"
uvicorn = {extras = ["standard"], version = "^0.41.0"}
sqlalchemy = "^2.0.48"
psycopg2-binary = "^2.9.10"
alembic = "^1.18.4"
pydantic-settings = "^2.13.1"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
redis = {extras = ["hiredis"], version = "^7.2.1"}
python-multipart = "^0.0.22"
httpx = "^0.28.1"
celery = "^5.6.2"
email-validator = "^2.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pymupdf (>=1.27.1,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "ruff (>=0.15.5,<0.16.0)",
    "black (>=26.1.0,<27.0.0)"
]
